+++ uCOS-II/os_cfg.h
@@ -138,7 +138,7 @@
 #define OS_Q_ACCEPT_EN            1u   /*     Include code for OSQAccept()                             */
 #define OS_Q_DEL_EN               1u   /*     Include code for OSQDel()                                */
 #define OS_Q_FLUSH_EN             1u   /*     Include code for OSQFlush()                              */
-#define OS_Q_PEND_ABORT_EN        0u   /*     Include code for OSQPendAbort()                          */
+#define OS_Q_PEND_ABORT_EN        1u   /*     Include code for OSQPendAbort()                          */
 #define OS_Q_POST_EN              1u   /*     Include code for OSQPost()                               */
 #define OS_Q_POST_FRONT_EN        1u   /*     Include code for OSQPostFront()                          */
 #define OS_Q_POST_OPT_EN          1u   /*     Include code for OSQPostOpt()                            */